backend/requirements.txt
Flask==3.0.3
openai==1.37.0
pandas==2.0.3
psycopg2_binary==2.9.9
Requests==2.32.3
